added predict command
