+++ ZkLobbyServer/ZkLobbyServer.csproj
@@ -87,6 +87,7 @@
     <Compile Include="autohost\Commands\CmdGame.cs" />
     <Compile Include="autohost\Commands\CmdMap.cs" />
     <Compile Include="autohost\Commands\CmdMaxPlayers.cs" />
+    <Compile Include="autohost\Commands\CmdPredict.cs" />
     <Compile Include="autohost\Commands\CmdRing.cs" />
     <Compile Include="autohost\Commands\CmdTitle.cs" />
     <Compile Include="autohost\Commands\CmdType.cs" />
